Update default image to fix RCE
Alpine 3.7 is vulnerable to RCE from a man-in-the-middle as described
here: https://justi.cz/security/2018/09/13/alpine-apk-rce.html

Update to Alpine 3.8.1, which fixes the issue. Alpine Docker does not
release tags for point releases, but you can verify the image is
updated correctly by running

    docker run alpine:3.8 cat /etc/alpine-release

which returns 3.8.1 for a fresh release.

Fixes #6.